Fix some UBSAN warnings with extended settings

+++ src/libdrakvuf/vmi.c
@@ -1922,7 +1922,7 @@ bool init_vmi(drakvuf_t drakvuf, bool fast_singlestep)
     }
 
     /* domain->max_pages is mostly just an annoyance that we can safely ignore */
-    rc = xen_set_maxmemkb(drakvuf->xen, drakvuf->domID, ~0);
+    rc = xen_set_maxmemkb(drakvuf->xen, drakvuf->domID, (uint64_t)~0);
     PRINT_DEBUG("Max mem set? %i\n", rc);
     if (rc < 0)
         return 0;
